Move the edits logic to the shared features namespaces,
as this will be used by code-formatting, code actions (quick fix),
and other features.
