+++ src/adapters/codemirror/feature.ts
@@ -18,6 +18,19 @@ export enum CommandEntryPoint {
   FileEditorContextMenu
 }
 
+function toDocumentChanges(changes: {
+  [uri: string]: lsProtocol.TextEdit[];
+}): lsProtocol.TextDocumentEdit[] {
+  let documentChanges = [];
+  for (let uri of Object.keys(changes)) {
+    documentChanges.push({
+      textDocument: { uri },
+      edits: changes[uri]
+    } as lsProtocol.TextDocumentEdit);
+  }
+  return documentChanges;
+}
+
 export interface IFeatureCommand {
   /**
    * The command id; it will be prepended with 'lsp' prefix.
@@ -215,5 +228,41 @@ export class CodeMirrorLSPFeature implements ILSPFeature {
     return range.editor
       .getDoc()
       .markText(range.start, range.end, { className: class_name });
+  }
+
+  protected apply_edit(workspaceEdit: lsProtocol.WorkspaceEdit) {
+    console.log(workspaceEdit);
+    let current_uri = this.connection.getDocumentUri();
+    // Specs: documentChanges are preferred over changes
+    let changes = workspaceEdit.documentChanges
+      ? workspaceEdit.documentChanges.map(
+          change => change as lsProtocol.TextDocumentEdit
+        )
+      : toDocumentChanges(workspaceEdit.changes);
+    for (let change of changes) {
+      let uri = change.textDocument.uri;
+      if (uri !== current_uri) {
+        console.warn('Workspace-wide edits not implemented yet');
+      } else {
+        // TODO: show "Renamed X to Y in {change.edits.length} places" in statusbar;
+        for (let edit of change.edits) {
+          let start = PositionConverter.lsp_to_cm(edit.range.start);
+          let end = PositionConverter.lsp_to_cm(edit.range.end);
+
+          let start_editor = this.virtual_document.get_editor_at_virtual_line(
+            start as IVirtualPosition
+          );
+          let end_editor = this.virtual_document.get_editor_at_virtual_line(
+            end as IVirtualPosition
+          );
+          if (start_editor !== end_editor) {
+            console.log('Edits not implemented for notebooks yet');
+          } else {
+            let doc = start_editor.getDoc();
+            doc.replaceRange(edit.newText, start, end);
+          }
+        }
+      }
+    }
   }
 }
